backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyMySQL==1.1.0
minio==7.2.0
pycryptodome==3.19.0
requests==2.31.0
beautifulsoup4==4.12.2
APScheduler==3.10.4 